+++ src/services/transcription/whisper_client.cpp
@@ -8,7 +8,7 @@
 #include "../../services/api/api.h"
 #include "../../screens/screen_transcription.h"
 #include <main.h>
-#include "../../screens/screen_display_extract.h"
+#include "../../screens/extracts/screen_food_extract.h"
 #include "../api/food_records.h"
 #include "../api/bath_records.h"
 #include <WiFiClientSecure.h>
